Fix data file locations and import/exports
