+++ src/data/users.js
@@ -1,4 +1,4 @@
-const usersData = [
+export const usersData = [
     {
       "name": "Saige O'Kon",
       "id": 1,
