--- test/test-ingredient-data.js
@@ -1,13 +1,0 @@
-// const ingredient1 = {"id": 20081, "name": "wheat flour","estimatedCostInCents": 142};
-// const ingredient2 = {"id": 18372,"name": "bicarbonate of soda", "estimatedCostInCents": 582};
-// const ingredient3 = {"id": 12345,"name": "fake crab", "estimatedCostInCents": 213};
-// const ingredient4 = {"id": 54321,"name": "ritz crackers", "estimatedCostInCents": 300};
-
-export const ingredientTestData = [
-  {"id": 20081, "name": "wheat flour","estimatedCostInCents": 142},
-  {"id": 18372,"name": "bicarbonate of soda", "estimatedCostInCents": 582},
-  {"id": 12345,"name": "fake crab", "estimatedCostInCents": 213},
-  {"id": 54321,"name": "ritz crackers", "estimatedCostInCents": 300}
-];
-
-// export default ingredientTestData;
