--- test/test-recipe-data.js
@@ -1,56 +1,0 @@
-export const recipeTestData = [
-  {
-    'id': 595736,
-    'image': 'potato.org',
-    'ingredients': [
-      {
-        "id": 20081,
-        "quantity": {
-          "amount": 2,
-          "unit": "c"
-        }
-      },
-      {
-        "id": 18372,
-        "quantity": {
-          "amount": 1,
-          "unit": "tsp"
-        }
-      }
-    ],
-    'instructions': [
-      {"instruction": "Preheat oven to 170 – 200°F", "number": 1},
-      {"instruction": "Mix with warm water", "number": 2}
-    ],
-    'name': 'Whole Grain Bread',
-    'tags': ['bread']
-  },
-  {
-    'id': 98765,
-    'image': 'crablover.org',
-    'ingredients': [
-      {
-        "id": 12345,
-        "quantity": {
-          "amount": 1,
-          "unit": "c"
-        }
-      },
-      {
-        "id": 54321,
-        "quantity": {
-        "amount": 4,
-        "unit": "crackers"
-        }
-      }
-    ],
-    'instructions': [
-      {"instruction": "Open package", "number": 1},
-      {"instruction": "combine", "number": 2}
-    ],
-    'name': 'Crabs on Crackers',
-    'tags': ["Hors d\'oeuvre", "dinner", "crab"]
-}
-];
-
-// export default recipeTestData;
